mylang: Pass a parent bblock to a build_typedef func
A parent bblock is a bblock in which all functions created during
folding has to be inserted. For the p_exec_start it is actually
the block that is popped in the begging of the function. So that
the problem with None parent of the p.parser.bblock is no longer
exists.