parse: Rewrite groupby_name function
The previous version of the function was a python itertools.groupby
wrapper. Consequently, it should receive sorted data. We have to
compute expressions in the natural order, so we cannot sacrifice the
bindings order.
As a result, the new version of the function groups bindings by tier
name as before, but in scope of its group the order of the bindings
remains initial.