parse: Fix expression computing order
In order to compute values of namespaces in the correct order
binding functions are computed right in the folding function.

So, previously:
1)  x: {
        a.b.c: 1
        a.d : 2
        a.b.f: 3
    }
=>

	folding_function():
	    x=ns(
	        a=ns(
	            b=ns(
	                c=1,
	                f=3
	                ),
	            d=2)
	        )

    Output:
    1
    3
    2

2)
    x: {
        a.b.c: 1
        a.d : 2
        a.b.f: 3
    }
=>

	folding_function():
		_c = c = 1
		_d = d = 2
		_f = f = 3
	    x=ns(
	        a=ns(
	            b=ns(
	                c=_c,
	                f=_f
	                ),
	            d=_d)
	        )

    Output:
    1
    2
    3

However, there remains one problem: the order is correct only
in scope of its outermost namespace.
In this example (without x) values are still computed in the incorrect
order:

a.b.c: 1
a.d : 2
a.b.f: 3

Output:
1
3
2