rgraph: Refactor rgraph itself
NodeContainer is devided into two classes: Rnode and Container. The
first one contains beacuseof ditionary with container as a key and
reason as a value and dictionry of its containers.
Container has set of its rnode members and therefore dictionary where
rnode is a key and reason is a value.

The rest of refactoring is about workaround and docs.