util: prop: Handle inheritance in 'cached_class_property' properly
Before this change behavior of caching a result of a getter depended
on whether a property was accessed on the class owning the property
or some of its subclasses, and even on the order of such accessing.

Now a 'cached_class_property' descriptor manages an internal mapping
that holds cached values, per each class from the subclass hierarchy.

There's also 'lazy_const_class_property' that saves a computed result
in the __dict__ of the class owning the property, replacing the latter
completely.