pydsl: (pypy) Avoid check against type(object.__init__)
This fixes a bug showing on pypy3: modules defined through `@module`
used to miss their options. The root cause is that `object.__init__` is
an ordinal function in PyPy, unlike CPython where it is a slot wrapper.
That is, user-defined methods have the same type as built-in ones in
PyPy.

Instead of checking types, just 'try'. IOW, LBYL -> EAFP :)